Added possibility to update books.

+++ Library/Library/MainWindows/BooksWindow.xaml.cs
@@ -6,6 +6,7 @@ using System.Windows;
 using System.Windows.Input;
 using Domain.Models;
 using Domain.Repository;
+using Library.AddWindows;
 using Library.Helpers;
 using Library.ViewModels;
 
@@ -91,7 +92,7 @@ namespace Library.MainWindows
 
         private void editCommand_Executed(object sender, ExecutedRoutedEventArgs e)
         {
-            DeleteBook();
+            EditBook();
         }
         private void deleteCommand_CanExecute(object sender, CanExecuteRoutedEventArgs e)
         {
@@ -104,12 +105,22 @@ namespace Library.MainWindows
 
         private void deleteCommand_Executed(object sender, ExecutedRoutedEventArgs e)
         {
-            DeleteBook();
+            MessageBoxResult messageBoxResult = System.Windows.MessageBox.Show("Are you sure?", "Delete Confirmation",
+                System.Windows.MessageBoxButton.YesNo);
+            if (messageBoxResult == MessageBoxResult.Yes)
+            {
+                DeleteBook();
+            }
         }
 
         public void EditBook()
         {
-            
+            var bookModel = booksTable.SelectedItem as BookModel;
+            if (bookModel != null)
+            {
+                var bookCreatingWindow = new AddWindows.AddNewBookWindow(_unitOfWork, _booksDisplayViewModel, bookModel);
+                bookCreatingWindow.ShowDialog();
+            }
         }
 
         public void DeleteBook()
